refactor: select root component
