--- src/components/Select/components/SelectTrigger/index.ts
@@ -1,9 +1,0 @@
-import { SelectTriggerButton } from './SelectTriggerButton';
-import { SelectTriggerField } from './SelectTriggerField';
-import { SelectTriggerRoot } from './SelectTriggerRoot';
-
-export const SelectTrigger = {
-    Root: SelectTriggerRoot,
-    Field: SelectTriggerField,
-    Button: SelectTriggerButton,
-};
