+++ src/components/Select/index.ts
@@ -1,11 +1,1 @@
-import { SelectItem } from './components/SelectItem';
-import { SelectList } from './components/SelectList';
-import { SelectRoot } from './components/SelectRoot';
-export * from './components/SelectTrigger';
-export * from './components/SelectRoot';
-
-export const Select = {
-    Root: SelectRoot,
-    List: SelectList,
-    Item: SelectItem,
-};
+export * from './components';
