+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export * from './Button';
-export * from './Select';
 export * from './Label';
-export * from './Select';
 export * from './Input';
 export * from './Feedback';
+export * from './Select';
